// src/components/AccountModals/LoginModal/login-modal-types.ts
enum LoginCase {
    lock = 'lock',
    mytables = 'mytables',
    logo = 'logo',
    link = 'link',
    share = 'share',
    error = 'error',
    default = 'default',
}

enum ModalType {
    LoginOptionsModal = 'LoginOptionsModal',
    LoginEmailModal = 'LoginEmailModal',
    SignupModal = 'SignupModal',
    ResetPasswordModal = 'ResetPasswordModal',
    EmailSentModal = 'EmailSentModal',
}

export { LoginCase, ModalType }